moved user profile types to admin client
