+++ js/libs/keycloak-admin-client/src/defs/userProfileMetadata.ts
@@ -1,10 +1,37 @@
-// See: https://github.com/keycloak/keycloak/blob/main/core/src/main/java/org/keycloak/representations/idm/UserProfileMetadata.java
-export interface UserProfileMetadata {
-  attributes?: UserProfileAttributeMetadata[];
-  groups?: UserProfileAttributeGroupMetadata[];
+export default interface UserProfileConfig {
+  attributes?: UserProfileAttribute[];
+  groups?: UserProfileGroup[];
+}
+export interface UserProfileAttribute {
+  name?: string;
+  validations?: Record<string, unknown>;
+  validators?: Record<string, unknown>;
+  annotations?: Record<string, unknown>;
+  required?: UserProfileAttributeRequired;
+  readOnly?: boolean;
+  permissions?: UserProfileAttributePermissions;
+  selector?: UserProfileAttributeSelector;
+  displayName?: string;
+  group?: string;
+}
+export interface UserProfileAttributeRequired {
+  roles?: string[];
+  scopes?: string[];
+}
+export interface UserProfileAttributePermissions {
+  view?: string[];
+  edit?: string[];
+}
+export interface UserProfileAttributeSelector {
+  scopes?: string[];
+}
+export interface UserProfileGroup {
+  name?: string;
+  displayHeader?: string;
+  displayDescription?: string;
+  annotations?: Record<string, unknown>;
 }
 
-// See: https://github.com/keycloak/keycloak/blob/main/core/src/main/java/org/keycloak/representations/idm/UserProfileAttributeMetadata.java
 export interface UserProfileAttributeMetadata {
   name?: string;
   displayName?: string;
@@ -15,10 +42,14 @@ export interface UserProfileAttributeMetadata {
   validators?: Record<string, Record<string, unknown>>;
 }
 
-// See: https://github.com/keycloak/keycloak/blob/main/core/src/main/java/org/keycloak/representations/idm/UserProfileAttributeGroupMetadata.java
 export interface UserProfileAttributeGroupMetadata {
   name?: string;
   displayHeader?: string;
   displayDescription?: string;
   annotations?: Record<string, unknown>;
+}
+
+export interface UserProfileMetadata {
+  attributes?: UserProfileAttributeMetadata[];
+  groups?: UserProfileAttributeGroupMetadata[];
 }
