+++ js/libs/keycloak-admin-client/src/resources/users.ts
@@ -7,7 +7,7 @@ import type { RequiredActionAlias } from "../defs/requiredActionProviderRepresen
 import type RoleRepresentation from "../defs/roleRepresentation.js";
 import type { RoleMappingPayload } from "../defs/roleRepresentation.js";
 import type UserConsentRepresentation from "../defs/userConsentRepresentation.js";
-import type { UserProfileConfig } from "../defs/userProfileConfig.js";
+import type UserProfileConfig from "../defs/userProfileMetadata.js";
 import type { UserProfileMetadata } from "../defs/userProfileMetadata.js";
 import type UserRepresentation from "../defs/userRepresentation.js";
 import type UserSessionRepresentation from "../defs/userSessionRepresentation.js";
