+++ js/libs/ui-shared/src/main.ts
@@ -19,5 +19,6 @@ export {
   setUserProfileServerError,
   isUserProfileError,
 } from "./user-profile/utils";
+export type { UserFormFields } from "./user-profile/utils";
 export { ScrollForm, mainPageContentId } from "./scroll-form/ScrollForm";
 export { FormPanel } from "./scroll-form/FormPanel";
